Update auth windows
Регистрация и вход теперь выполняется в разных окнах.
Изменины сами окна.
