+++ app/src/main/AndroidManifest.xml
@@ -9,10 +9,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <activity android:name=".UI.auth.RegistrationApp"></activity>
         <activity
             android:name=".UI.task_number_6.Activity_Task_06"
             android:label="@string/title_activity_task_06"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
+            android:theme="@style/AppTheme.NoActionBar" />
         <activity
             android:name=".UI.task_number_5.Activity_Task_05"
             android:label="@string/title_activity_task_05"
@@ -28,11 +29,9 @@
         <activity
             android:name=".Activities.Activity_Main"
             android:label="@string/app_name"
-            android:theme="@style/AppTheme.NoActionBar">
-
-        </activity>
+            android:theme="@style/AppTheme.NoActionBar"></activity>
         <activity
-            android:name=".UI.register.reg"
+            android:name=".UI.auth.SignInApp"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -41,6 +40,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
