+++ app/src/main/res/drawable/btn_register.xml
@@ -8,7 +8,7 @@
             <solid android:color="@color/btn_register_stroke"/>
             <stroke android:color="@color/btn_register"
                 android:width="2dp"/>
-            <corners android:radius="2dp"/>
+            <corners android:radius="30dp"/>
         </shape>
     </item>
