+++ app/src/main/res/layout/fragment_task_0513.xml
@@ -15,6 +15,7 @@
         android:layout_marginTop="16dp"
         android:text="Введите количество букв:"
         android:textColor="#2B2424"
+        android:textColorHint="#FFFFFF"
         android:textSize="18sp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
