--- app/src/main/res/layout/register_window.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent"
-
-    app:cardElevation="10dp">
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="20dp"
-        android:orientation="vertical">
-
-        <com.rengwuxian.materialedittext.MaterialEditText
-            android:id="@+id/email_field"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Email"
-            android:text=""
-            android:textColorHint="@color/colorPrimaryDark"
-            android:textColor="@color/colorPrimary"
-            app:met_floatingLabel="highlight"
-            android:inputType="textEmailAddress"
-            app:met_baseColor="#004BB1"
-            app:met_primaryColor="#982360"
-            app:met_errorColor="#FF0000"
-            app:met_singleLineEllipsis="true"/>
-
-        <com.rengwuxian.materialedittext.MaterialEditText
-            android:id="@+id/login_field"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Login"
-            android:inputType="textEmailAddress"
-            android:text=""
-            android:textColor="@color/colorPrimary"
-            android:textColorHint="@color/colorPrimaryDark"
-            app:met_baseColor="#004BB1"
-            app:met_errorColor="#FF0000"
-            app:met_floatingLabel="highlight"
-            app:met_primaryColor="#982360"
-            app:met_singleLineEllipsis="true" />
-
-        <com.rengwuxian.materialedittext.MaterialEditText
-            android:id="@+id/password_field"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Password"
-            android:text=""
-            android:textColorHint="@color/colorPrimaryDark"
-            android:textColor="@color/colorPrimary"
-            app:met_floatingLabel="highlight"
-            android:inputType="textPassword"
-            app:met_baseColor="#004BB1"
-            app:met_primaryColor="#982360"
-            app:met_errorColor="#FF0000"
-            app:met_singleLineEllipsis="true"/>
-
-    </LinearLayout>
-
-</androidx.cardview.widget.CardView>
