+++ app/src/main/res/layout/registration_window.xml
@@ -6,7 +6,8 @@
     style="@style/Theme.AppCompat.DayNight.NoActionBar"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.auth.SignInApp">
+    tools:context=".UI.auth.RegistrationApp"
+    >
 
 
     <LinearLayout
@@ -46,7 +47,7 @@
 
 
         <com.rengwuxian.materialedittext.MaterialEditText
-            android:id="@+id/sign_in_login"
+            android:id="@+id/registration_login"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="15dp"
@@ -54,12 +55,22 @@
             android:inputType="textEmailAddress" />
 
         <com.rengwuxian.materialedittext.MaterialEditText
-            android:id="@+id/sign_in_password"
+            android:id="@+id/registration_mail"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_margin="15dp"
+            android:hint="Login"
+            android:inputType="textPersonName"/>
+
+        <com.rengwuxian.materialedittext.MaterialEditText
+            android:id="@+id/registration_password"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="15dp"
             android:hint="Password"
-            android:inputType="textEmailAddress" />
+            android:inputType="textPassword" />
+
+
     </LinearLayout>
 
     <LinearLayout
@@ -69,44 +80,26 @@
         android:layout_alignParentBottom="true"
         android:layout_marginLeft="30dp"
         android:layout_marginRight="30dp"
-        android:layout_marginBottom="10dp"
+        android:layout_marginBottom="25dp"
         android:gravity="center"
         android:orientation="vertical"
         android:weightSum="2">
 
 
         <Button
-            android:id="@+id/btnSignIn"
+            android:id="@+id/btnRegistration"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginBottom="6dp"
             android:layout_weight="1"
             android:background="@drawable/btn_sign_in"
             android:fontFamily="@font/montserrat"
-            android:text="Войти"
-            android:textColor="@color/white" />
-
-        <Button
-            android:id="@+id/sing_in_with_others"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="6dp"
-            android:layout_weight="1"
-            android:background="@drawable/btn_register"
-            android:fontFamily="@font/montserrat"
-            android:text="Войти с помощью"
-            android:textColor="@color/white" />
-
-        <TextView
-            android:id="@+id/tv_register"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:clickable="true"
             android:text="Зарегистрироваться"
-            android:textAlignment="center"
-            android:textAllCaps="false"
-            android:textStyle="italic" />
+            android:textColor="@color/white" />
+
+
+
+
     </LinearLayout>
 
 </RelativeLayout>
